requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
Django==1.9.2
djangorestframework==3.3.2
idna==2.6
pyfcm==1.4.2
pytz==2015.7
requests==2.18.4
requests-toolbelt==0.8.0
urllib3==1.22
